@@ -76,6 +76,7 @@
     "charenc": "^0.0.2",
     "crypt": "^0.0.2",
     "crypto-js": "^4.2.0",
+    "framer-motion": "^12.12.1",
     "fuse.js": "^6.6.2",
     "immer": "^9.0.12",
     "is-buffer": "~1.1.6",
@@ -83,8 +84,8 @@
     "lowlight": "^2.8.1",
     "nodemailer": "^6.9.1",
     "pg": "^8.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-error-boundary": "^3.1.4",
     "react-helmet-async": "^1.2.3",
     "react-hot-toast": "^2.4.0",
@@ -145,8 +146,8 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.23",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-redux": "^7.1.23",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-transition-group": "^4.4.4",
