@@ -28,4 +28,4 @@ function Footer() {
   );
 }
 
-export default Footer;
+export default Footer; 
