@@ -23,7 +23,7 @@ function AppWrapper(props) {
               sidebarIsOpen
                 ? "w-full sm:w-64 px-4 pb-4 pt-16 sm:pt-4 absolute sm:relative top-0 bottom-0 left-0 right-0 bg-white z-20"
                 : "w-64 p-4 hidden sm:visible sm:flex"
-            } border-r-2 border-gray-200 flex flex-col max-h-screen shrink-0 bg-white`}
+            } border-r-2 border-gray-100 flex flex-col max-h-screen shrink-0 bg-white`}
           >
             <Sidebar {...props} closeSidebar={() => setSidebarIsOpen(false)} />
           </div>
