@@ -46,4 +46,4 @@ function NotFound() {
   );
 }
 
-export default NotFound;
+export default NotFound; 
