@@ -3,7 +3,7 @@ import PublicNoteContainer from '../containers/PublicNoteContainer';
 
 function PublicNote() {
   return (
-    <div className="w-full h-full min-h-full">
+    <div className="w-full h-full min-h-full flex-grow flex flex-col">
       <PublicNoteContainer />
     </div>
   );
